+++ CHANGELOG.md
@@ -44,7 +44,8 @@
 -   TypeDoc will now produce more informative error messages for options which cannot be set from the cli, #2022.
 -   TypeDoc will now attempt to guess what option you may have meant if given an invalid option name.
 -   TypeDoc options may now be set under the `typedocOptions` key in `package.json`, #2112.
--   Moved sidebar to left of content for consistency with most other websites, #2189
+-   Moved sidebar to left of content for consistency with most other websites, #2189.
+-   Added `--cacheBust` option to tell TypeDoc to include include the generation time in files, #2124.
 
 ### Bug Fixes